# Hyphenation: speed up book loading by setting hyph algo earlier