@poire-z 2018-08-07
Well, this speedup is noticable mostly on the **first** opening of big books. On next openings, if a cache is re-used, this late hyphenation wouldn't cause any problem, and the already fast loading won't get any faster.
So, it needs a big book to be noticable, but as big books make a cache... we won't notice it often :)
But well, I benchmark on first opening, so that suits me.