@robert00s 2018-08-07
Every speedup is worthy of praise :)